Print a 'sbt_std_boot' Object

Print method for a 'sbt_std_boot' object, which is the output of standardizedSolution_boot().

Usage

# S3 method for class 'sbt_std_boot'
print(
  x,
  ...,
  nd = 3,
  output = c("lavaan.printer", "text", "table"),
  standardized_only = TRUE,
  boot_ci_only = FALSE,
  drop_cols = "Z"
)

Arguments

x

Object of the class sbt_std_boot, the output of standardizedSolution_boot().

...

Optional arguments to be passed to print() methods.

nd

The number of digits after the decimal place. Default is 3.

output

String. How the results are printed. If set to "table", the results are printed in a table format similar to that of lavaan::standardizedSolution(). If set to "text", the results will be printed in a text format similar to the printout of the output of summary() of a 'lavaan'-class object. If set to "lavaan.printer", the default, lavaan.printer will be used to print a more compact version of the "text" output.

standardized_only

Logical. If TRUE, the default, only the results for the standardized solution will be printed. If FALSE, then the standardized solution is printed alongside the unstandardized solution, as in the printout of the output of summary() of a 'lavaan'-class object.

boot_ci_only

Logical. Whether only bootstrap confidence intervals are printed. If FALSE, the default, the delta method confidence intervals by lavaan::standardizedSolution() are also printed.

drop_cols

The name(s) of the column(s) to drop if output format is "lavaan.printer". Default is "Z", to fit the print out to the usual screen width of 80.

Value

x is returned invisibly. Called for its side effect.

Details

The default format of the printout, "lavaan.printer", is a compact version of the lavaan-style printout, generated by lavaan.printer. Alternatively, users can request a format similar to that of the printout of the summary of a lavaan output by setting output to "text". This format can be used if "lavaan.printer" failed.

Users can also print the content just as a data frame by setting output to "table". Not easy to read much more compact.

For the "text" or "lavaan.printer" format, users can also select whether only the standardized solution is printed (the default) or whether the standardized solution is appended to the right of the printout.

See also

Author

Shu Fai Cheung https://orcid.org/0000-0002-9871-9448

Examples

library(lavaan)
set.seed(5478374)
n <- 50
x <- runif(n) - .5
m <- .40 * x + rnorm(n, 0, sqrt(1 - .40))
y <- .30 * m + rnorm(n, 0, sqrt(1 - .30))
dat <- data.frame(x = x, y = y, m = m)
model <-
'
m ~ a*x
y ~ b*m
ab := a*b
'

# Should set bootstrap to at least 2000 in real studies
fit <- sem(model, data = dat, fixed.x = FALSE,
           se = "boot",
           bootstrap = 50)
std_out <- standardizedSolution_boot(fit)
#> Warning: The number of bootstrap samples (50) is less than 'boot_pvalue_min_size' (1000). Bootstrap p-values are not computed.
std_out
#> 
#> Bootstrapping:
#>                                     
#>  Valid Bootstrap Samples: 50        
#>  Level of Confidence:     95.0%     
#>  CI Type:                 Percentile
#>  Standardization Type:    std.all   
#> 
#> Parameter Estimates Settings:
#>                                                 
#>  Standard errors:                      Bootstrap
#>  Number of requested bootstrap draws:  50       
#>  Number of successful bootstrap draws: 50       
#> 
#> Regressions:
#>            Std    SE     p  CI.Lo CI.Up   bSE bCI.Lo bCI.Up
#>  m ~                                                       
#>   x (a)  0.229 0.117 0.051 -0.001 0.458 0.121 -0.025  0.462
#>  y ~                                                       
#>   m (b)  0.198 0.121 0.100 -0.038 0.434 0.112 -0.005  0.492
#> 
#> Variances:
#>            Std    SE     p  CI.Lo CI.Up   bSE bCI.Lo bCI.Up
#>   .m     0.948 0.053 0.000  0.843 1.053 0.055  0.786  1.000
#>   .y     0.961 0.048 0.000  0.867 1.054 0.056  0.758  0.999
#>    x     1.000                                             
#> 
#> Defined Parameters:
#>            Std    SE     p  CI.Lo CI.Up   bSE bCI.Lo bCI.Up
#>  ab (ab) 0.045 0.037 0.215 -0.026 0.117 0.038 -0.006  0.151
#> 
#> Footnote:
#> - Std: Standardized estimates.
#> - SE: Delta method standard errors.
#> - p: Delta method p-values.
#> - CI.Lo, CI.Up: Delta method confidence intervals.
#> - bSE: Bootstrap standard errors.
#> - bCI.Lo, bCI.Up: Bootstrap confidence intervals.
print(std_out, standardized_only = FALSE)
#> 
#> Bootstrapping:
#>                                     
#>  Valid Bootstrap Samples: 50        
#>  Level of Confidence:     95.0%     
#>  CI Type:                 Percentile
#>  Standardization Type:    std.all   
#> 
#> Parameter Estimates Settings:
#>                                                 
#>  Standard errors:                      Bootstrap
#>  Number of requested bootstrap draws:  50       
#>  Number of successful bootstrap draws: 50       
#> 
#> Regressions:
#>            Est    SE     p  CI.Lo CI.Up   Std   bSE bCI.Lo bCI.Up
#>  m ~                                                             
#>   x (a)  0.569 0.293 0.052 -0.056 1.165 0.229 0.121 -0.025  0.462
#>  y ~                                                             
#>   m (b)  0.219 0.147 0.136 -0.002 0.725 0.198 0.112 -0.005  0.492
#> 
#> Variances:
#>            Est    SE     p  CI.Lo CI.Up   Std   bSE bCI.Lo bCI.Up
#>   .m     0.460 0.083 0.000  0.247 0.593 0.948 0.055  0.786  1.000
#>   .y     0.570 0.109 0.000  0.341 0.792 0.961 0.056  0.758  0.999
#>    x     0.078 0.013 0.000  0.052 0.102 1.000                    
#> 
#> Defined Parameters:
#>            Est    SE     p  CI.Lo CI.Up   Std   bSE bCI.Lo bCI.Up
#>  ab (ab) 0.125 0.107 0.246 -0.019 0.440 0.045 0.038 -0.006  0.151
#> 
#> Footnote:
#> - Estimate: Unstandardized estimates.
#> - SE: Standard errors of unstandardized estimates.
#> - p: P-values of unstandardized estimates.
#> - CI.Lo, CI.Up: Confidence intervals of unstandardized estimates.
#> - Std: Standardized estimates.
#> - bSE: Bootstrap standard errors of standardized estimates.
#> - bCI.Lo, bCI.Up: Bootstrap confidence intervals of standardized
#>   estimates.